[core] Rename interal folder internals on x-date-picker-pro
