+++ packages/x-date-pickers-pro/src/DesktopDateRangePicker/DesktopDateRangePicker.types.ts
@@ -3,7 +3,7 @@ import {
   UseDesktopRangePickerSlotsComponent,
   UseDesktopRangePickerSlotsComponentsProps,
   DesktopRangeOnlyPickerProps,
-} from '../internal/hooks/useDesktopRangePicker';
+} from '../internals/hooks/useDesktopRangePicker';
 import {
   BaseDateRangePickerProps,
   BaseDateRangePickerSlotsComponent,
