+++ packages/x-date-pickers-pro/src/MultiInputDateTimeRangeField/index.ts
@@ -1,5 +1,5 @@
 export { MultiInputDateTimeRangeField } from './MultiInputDateTimeRangeField';
-export { useMultiInputDateTimeRangeField as unstable_useMultiInputDateTimeRangeField } from '../internal/hooks/useMultiInputRangeField/useMultiInputDateTimeRangeField';
+export { useMultiInputDateTimeRangeField as unstable_useMultiInputDateTimeRangeField } from '../internals/hooks/useMultiInputRangeField/useMultiInputDateTimeRangeField';
 export type {
   UseMultiInputDateTimeRangeFieldProps,
   UseMultiInputDateTimeRangeFieldComponentProps,
