+++ packages/x-date-pickers-pro/src/MultiInputTimeRangeField/index.ts
@@ -1,5 +1,5 @@
 export { MultiInputTimeRangeField } from './MultiInputTimeRangeField';
-export { useMultiInputTimeRangeField as unstable_useMultiInputTimeRangeField } from '../internal/hooks/useMultiInputRangeField/useMultiInputTimeRangeField';
+export { useMultiInputTimeRangeField as unstable_useMultiInputTimeRangeField } from '../internals/hooks/useMultiInputRangeField/useMultiInputTimeRangeField';
 export type {
   UseMultiInputTimeRangeFieldProps,
   UseMultiInputTimeRangeFieldComponentProps,
