+++ packages/x-date-pickers-pro/src/internals/hooks/useDesktopRangePicker/index.ts
