+++ packages/x-date-pickers-pro/src/internals/hooks/useStaticRangePicker/index.ts
