+++ packages/x-date-pickers-pro/src/internals/models/range.ts
