+++ packages/x-date-pickers-pro/src/internals/models/rangePickerProps.ts
