+++ packages/x-date-pickers-pro/src/internals/utils/date-range-manager.ts
