+++ packages/x-date-pickers-pro/src/internals/utils/validation/validateDateRange.ts
